refactor: change files name and use try catch in controller
