+++ src/app/errors/AccessDeniedError.js
