+++ src/app/server.js
@@ -1,20 +1,20 @@
 /* importar o módulo do framework express */
-const express = require("express")
+const express = require('express')
 
 /* Necessário para acesso de outros domínos */
-var cors = require("cors")
+var cors = require('cors')
 
 /* Para configurar as variáveis de ambiente da aplicação */
-require("dotenv/config")
+require('dotenv/config')
 
 /* importar o módulo do consign */
-const consign = require("consign")
+const consign = require('consign')
 
 /* importar o módulo do body-parser */
-const bodyParser = require("body-parser")
+const bodyParser = require('body-parser')
 
 /* importar o módulo do express-validator */
-const session = require("express-session")
+const session = require('express-session')
 
 /* iniciar o objeto do express */
 const app = express()
@@ -28,7 +28,7 @@ app.use(bodyParser.json())
 
 /* Configuração da sessão */
 const configSession = {
-  secret: "carpssession", // Chave secreta
+  secret: 'carpssession', // Chave secreta
   resave: false, // Se true, a sessão será regravada no servidor
   saveUninitialized: false, // Se true cria uma nova sessão sempre que a mesma for modificada.
 }
@@ -43,39 +43,36 @@ app.use(cors())
 
 app.use(function (req, res, next) {
   /* Habilita requisições cros domain, requisições de domínos diferentes 
-  res.setHeader("Access-Control-Allow-Origin", "*")
+  res.setHeader('Access-Control-Allow-Origin', '*')
 
   */
   /* Quais os métodos que a origem pode requisitar 
-  res.setHeader("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE")
+  res.setHeader('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE')
   */
 
   /* Habilitar que a requisição feita pela origem tenha cabeçalhos reescritos 
   res.setHeader(
-    "Access-Control-Allow-Headers",
-    "Origin, X-Requested-With, Content-Type, Accept"
+    'Access-Control-Allow-Headers',
+    'Origin, X-Requested-With, Content-Type, Accept'
     )
   */
 
   /*  
-  res.setHeader("Access-Control-Allow-Credentials", true)
+  res.setHeader('Access-Control-Allow-Credentials', true)
   */
 
   /* Tratamento para acesso aos recuros
      Verifica se existe um usuário logado
-     Para login e logout veja no arquivo acessoRoute.js
+     Para login e logout veja no arquivo AcessoRoute.js
   */
-  if (process.env.NODE_ENV === "production") {
+  if (process.env.NODE_ENV === 'production') {
     let urlOrigem = req.originalUrl
-    if (urlOrigem === "/" || urlOrigem === "/login") {
+    if (urlOrigem === '/' || urlOrigem === '/login') {
       next()
       return
     }
     if (req.session.usuario === undefined) {
-      res.status(401).json({
-        mensagem:
-          "Não existe um usuário logado. Por favor tente fazer o login.",
-      })
+      res.status(401).json({mensagem: 'Não existe um usuário logado. Por favor tente fazer o login.'})
       return
     }
   }
@@ -83,30 +80,30 @@ app.use(function (req, res, next) {
 })
 
 /* efetua o autoload das rotas, dos models e dos controllers para o objeto app */
-const caminho = "./src/"
+const caminho = './src/'
 /* Configuração do consign para utilização dos arquivos sem a necessidade de chamar a função require() */
 consign({
-  locale: "pt-br",
+  locale: 'pt-br',
 })
-  .include(caminho + "db/models/index.js")
-  .then(caminho + "app/util")
-  .then(caminho + "db/repository")
-  .then(caminho + "controller")
-  .then(caminho + "app/route")
+  .include(caminho + 'db/models/index.js')
+  .then(caminho + 'app/util')
+  .then(caminho + 'db/repository')
+  .then(caminho + 'controller')
+  .then(caminho + 'app/route')
   .into(app)
 
 /* Configuração para endereços não existentes na aplicação */
 app.use(function (req, res, next) {
-  res.status(404).json({ mensagem: "Rota não encontrada." })
+  res.status(404).json({ mensagem: 'Rota não encontrada.' })
   next()
 })
 
 /* Configuração para erros internos da aplicação */
 app.use(function (err, req, res, next) {
   res.status(500).json({
-    mensagem: "Erro interno.",
+    mensagem: 'Erro interno.',
     error: err.toString(),
-    contato: "jairmaiag@gmail.com",
+    contato: 'jairmaiag@gmail.com',
   })
   next()
 })
