+++ src/controller/acesso/AcessoController.js
@@ -1,12 +1,28 @@
+const { serverError, badRequest, ok } = require('../../app/helpers/http/HttpHelpers')
+
 class AcessoController {
 
   constructor(app) {
     this.app = app
-    this.repository = new this.app.src.db.repository.UsuarioRepository(this.app)
+    this.repository = new this.app.src.db.repository.UsuarioRepository()
   }
 
-  async login(filter) {
-    return await this.repository.findByLoginSenha(filter)
+  async login(req) {
+    try {
+      const filter = req.body
+      if (!filter.login || !filter.senha) {
+        return badRequest('Campos de Login e Senha são obrigatórios.')
+      }
+
+      const entity = await this.repository.findByLoginSenha(filter)
+      if (!entity) {
+        return badRequest('Usuário ou senha inválida')
+      }
+
+      return ok(entity)
+    } catch (error) {
+      return serverError(error)
+    }
   }
 
   async logout() { }
@@ -15,10 +31,6 @@ class AcessoController {
 
 function retorno(app) {
   return new AcessoController(app)
-}
-
-module.exports = function () {
-  return retorno
 }
 
 module.exports = () => retorno
