+++ src/controller/produto/ProdutoController.js
@@ -1,12 +1,12 @@
 const { getUUIDV4 } = require('../../app/util/UUIDGenerator')
-const { serverError, ok, noContent, notFound, forbidden } = require('../../app/helpers/http/http-helpers')
-const InvalidParamError = require('../../app/errors/invalid-param-error')
+const { serverError, ok, noContent, notFound, forbidden } = require('../../app/helpers/http/HttpHelpers')
+const InvalidParamError = require('../../app/errors/InvalidParamError')
 
 class ProdutoController {
 
   constructor(app) {
     this.app = app
-    this.repository = new this.app.src.db.repository.ProdutoRepository(this.app)
+    this.repository = new this.app.src.db.repository.ProdutoRepository()
   }
 
   async index(req) {
@@ -14,7 +14,7 @@ class ProdutoController {
       const entities = await this.repository.findAll(req.body.attributes, req.body.filter, req.body.order)
       return entities.length > 0 ? ok(entities) : noContent()
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -23,7 +23,7 @@ class ProdutoController {
       const entities = await this.repository.findAndPaginate(req.body.attributes, req.body.filter, req.body.order, req.body.page)
       return entities.rows.length > 0 ? ok(entities) : noContent()
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -32,7 +32,7 @@ class ProdutoController {
       const entity = await this.repository.findByUUId(req.params.UUId)
       return entity ? ok(entity) : noContent()
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -41,7 +41,7 @@ class ProdutoController {
       const entity = await this.repository.findById(req.params.id)
       return entity ? ok(entity) : noContent()
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -55,7 +55,7 @@ class ProdutoController {
       const entity = await this.repository.insert(dados)
       return entity ? ok(entity) : notFound('Erro ao inserir o registro')
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -69,7 +69,7 @@ class ProdutoController {
       const entity = await this.repository.update(dados)
       return entity ? ok(entity) : notFound('Registro não encontrado')
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -78,7 +78,7 @@ class ProdutoController {
       const quantidadeDeletada = await this.repository.delete(req.params.UUId)
       return quantidadeDeletada > 0 ? ok(quantidadeDeletada) : notFound('Registro não encontrado')
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 }
