+++ src/controller/usuario/UsuarioController.js
@@ -1,12 +1,12 @@
-const InvalidParamError = require('../../app/errors/invalid-param-error')
+const InvalidParamError = require('../../app/errors/InvalidParamError')
 const { getUUIDV4 } = require('../../app/util/UUIDGenerator')
-const { serverError, ok, noContent, notFound, forbidden } = require('../../app/helpers/http/http-helpers')
+const { serverError, ok, noContent, notFound, forbidden } = require('../../app/helpers/http/HttpHelpers')
 
 class UsuarioController {
 
   constructor(app) {
     this.app = app
-    this.repository = new this.app.src.db.repository.UsuarioRepository(this.app)
+    this.repository = new this.app.src.db.repository.UsuarioRepository()
   }
 
   async index(req) {
@@ -14,7 +14,7 @@ class UsuarioController {
       const entities = await this.repository.findAll(req.body.attributes, req.body.filter, req.body.order)
       return entities.length > 0 ? ok(entities) : noContent()
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -23,7 +23,7 @@ class UsuarioController {
       const entities = await this.repository.findAndPaginate(req.body.attributes, req.body.filter, req.body.order, req.body.page)
       return entities.rows.length > 0 ? ok(entities) : noContent()
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -32,7 +32,7 @@ class UsuarioController {
       const entity = await this.repository.findByUUId(req.params.UUId)
       return entity ? ok(entity) : noContent()
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -41,7 +41,7 @@ class UsuarioController {
       const entity = await this.repository.findById(req.params.id)
       return entity ? ok(entity) : noContent()
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -79,7 +79,7 @@ class UsuarioController {
       const entity = await this.repository.insert(dados)
       return entity ? ok(entity) : notFound('Erro ao inserir o registro')
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -93,7 +93,7 @@ class UsuarioController {
       const entity = await this.repository.update(dados)
       return entity ? ok(entity) : notFound('Registro não encontrado')
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 
@@ -102,7 +102,7 @@ class UsuarioController {
       const quantidadeDeletada = await this.repository.delete(req.params.UUId)
       return quantidadeDeletada > 0 ? ok(quantidadeDeletada) : notFound('Registro não encontrado')
     } catch (error) {
-      serverError(error)
+      return serverError(error)
     }
   }
 }
