+++ src/db/models/index.js
@@ -5,7 +5,7 @@ const path = require("path")
 const Sequelize = require("sequelize")
 const basename = path.basename(__filename)
 const env = process.env.NODE_ENV || "development"
-const databaseConfig = require('../config/config')[env]
+const databaseConfig = require('../config/Config')[env]
 databaseConfig.username = process.env.DATABASEUSERNAME || databaseConfig.username
 databaseConfig.password = process.env.DATABASEPASSWORD || databaseConfig.password
 databaseConfig.database = process.env.DATABASENAME || databaseConfig.database
